Enable non-standard ports
Corrects the port variable in a couple of places to allow non-standard ports to work correctly. Before this change the replies are ignored as they are compared to the PORT constant instead of the port variable, and then the retry is sent to the PORT constant instead of the port variable, effectively blocking any attempts to use non-standard ports.